#3: sdds_files split into sdds and tbt, simplification in progress
